# faster array indexing via boolean masks. Addresses #412.