--- a/poppy/tests/test_optics.py
+++ b/poppy/tests/test_optics.py
@@ -628,16 +628,16 @@ def test_ThinLens(display=False):
 
     # Now test the values at some precisely chosen pixels
     y, x = wave.coordinates()
-    at_radius = np.where((x==1) & (y==0))
+    at_radius = ((x==1) & (y==0))
     assert np.allclose(wave.phase[at_radius], np.pi/2), "Didn't get 1/2 wave OPD at edge of optic"
     assert len(at_radius[0]) > 0, "Array indices messed up - need to have a pixel at exactly (1,0)"
 
-    at_radius = np.where((x==0) & (y==1))
+    at_radius = ((x==0) & (y==1))
     assert np.allclose(wave.phase[at_radius], np.pi/2), "Didn't get 1/2 wave OPD at edge of optic"
     assert len(at_radius[0]) > 0, "Array indices messed up - need to have a pixel at exactly (0,1)"
 
 
-    at_center = np.where((x==0) & (y==0))
+    at_center = ((x==0) & (y==0))
     assert np.allclose(wave.phase[at_center], -np.pi/2), "Didn't get -1/2 wave OPD at center of optic"
     assert len(at_radius[0]) > 0, "Array indices messed up - need to have a pixel at exactly (0,0)"