Low-level change in numpy array indexing syntax to use e.g. `array[blah is true]` rather than `w = np.where(blah is true); array[w]`. Speeds up a variety of tasks such as optic setup and sampling.  A small optimization for performance but adds up over time.

I didn't extensively benchmark every part individually, but some spot tests confirm noticeable speedups:

**current develop**:

```
%pylab inline
import poppy

ap = poppy.HexagonAperture()
%timeit ap.sample(npix=1024)
43.8 ms ± 1.51 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

ap = poppy.CircularAperture()
%timeit ap.sample(npix=1024)
45.5 ms ± 1.22 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

ap = poppy.SquareAperture()
%timeit ap.sample(npix=1024)
40 ms ± 1.28 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

**with this PR:**

```
ap = poppy.HexagonAperture()
%timeit ap.sample(npix=1024)
30.8 ms ± 768 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

ap = poppy.CircularAperture()
%timeit ap.sample(npix=1024)
40.6 ms ± 1.39 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

ap = poppy.SquareAperture()
%timeit ap.sample(npix=1024)
27.7 ms ± 1.3 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```